feat: introduce CheckGroupV2 which avoids unintuitive implicit defaults
The current CheckGroup continues to work the same as before, however a
warning is output recommending the user to upgrade to CheckGroupV2.
