+++ packages/cli/src/constructs/check-group.ts
@@ -1,538 +1,9 @@
-import * as path from 'path'
-import { glob } from 'glob'
-import { Ref } from './ref'
-import { Session } from './project'
-import { Construct } from './construct'
-import { BrowserCheck } from './browser-check'
-import { AlertChannel, AlertChannelRef } from './alert-channel'
-import { EnvironmentVariable } from './environment-variable'
-import { AlertChannelSubscription } from './alert-channel-subscription'
-import { PrivateLocation, PrivateLocationRef } from './private-location'
-import { PrivateLocationGroupAssignment } from './private-location-group-assignment'
-import { CheckConfigDefaults } from '../services/checkly-config-loader'
-import { ApiCheckDefaultConfig } from './api-check'
-import { pathToPosix } from '../services/util'
-import type { Region } from '..'
-import { type Frequency } from './frequency'
-import {
-  type RetryStrategy,
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  type RetryStrategyBuilder, // Used for @links in comments.
-} from './retry-strategy'
-import { AlertEscalation } from './alert-escalation-policy'
-import { MultiStepCheck } from './multi-step-check'
-import CheckTypes from '../constants'
-import { Diagnostics } from './diagnostics'
-import { DeprecatedPropertyDiagnostic, InvalidPropertyValueDiagnostic } from './construct-diagnostics'
+export * from './check-group-v1'
+export * from './check-group-v2'
+export * from './check-group-ref'
 
-const defaultApiCheckDefaults: ApiCheckDefaultConfig = {
-  headers: [],
-  queryParameters: [],
-  url: '',
-  basicAuth: {
-    username: '',
-    password: '',
-  },
-}
+export {
+  CheckGroupV1 as CheckGroup,
+  CheckGroupV1Props as CheckGroupProps,
+} from './check-group-v1'
 
-type BrowserCheckConfig = CheckConfigDefaults & {
-  /**
-   * Glob pattern to include multiple files, i.e. all `.spec.ts` files
-   */
-  testMatch: string | string[],
-}
-
-type MultiStepCheckConfig = CheckConfigDefaults & {
-  /**
-   * Glob pattern to include multiple files, i.e. all `.spec.ts` files
-   */
-  testMatch: string | string[],
-}
-
-export interface CheckGroupProps {
-  /**
-   * The name of the check group.
-   */
-  name: string
-
-  /**
-   * Determines whether the checks in the group are running or not.
-   *
-   * When `true`, all checks in the group that are also activated will run.
-   *
-   * When `false`, no checks in the group will run, regardless of whether they
-   * are activated or not.
-   *
-   * If not set, individual check settings are used.
-   */
-  activated?: boolean
-
-  /**
-   * Determines if any notifications will be sent out when a check in this
-   * group fails and/or recovers.
-   *
-   * When `true`, all checks in the group act as if they are muted, regardless
-   * of their own state, and will not trigger alerts.
-   *
-   * When `false`, all checks in the group that are also not muted will trigger
-   * alerts.
-   *
-   * If not set, individual check settings are used.
-   */
-  muted?: boolean
-
-  /**
-   * Setting this to "true" will trigger a retry when a check fails from
-   * the failing region and another, randomly selected region before marking
-   * the check as failed.
-   *
-   * If set, overrides the doubleCheck property of all checks in the group.
-   *
-   * If not set, individual check settings are used.
-   *
-   * @deprecated Use {@link CheckGroupProps.retryStrategy} instead.
-   */
-  doubleCheck?: boolean
-
-  /**
-   * The runtime version, i.e. fixed set of runtime dependencies, used to
-   * execute checks in this group.
-   *
-   * This value is only used as a fallback for checks that do not explicitly
-   * choose a runtime.
-   *
-   * If not set, the ultimate fallback is the account default runtime.
-   */
-  runtimeId?: string
-
-  /**
-   * An array of one or more data center locations where to run the checks.
-   *
-   * If either {@link CheckGroupProps.locations} or
-   * {@link CheckGroupProps.privateLocations} is set to a non-empty value, all
-   * checks in the group will use those values instead of their own.
-   */
-  locations?: (keyof Region)[]
-
-  /**
-   * An array of one or more private locations where to run the checks.
-   *
-   * If either {@link CheckGroupProps.locations} or
-   * {@link CheckGroupProps.privateLocations} is set to a non-empty value, all
-   * checks in the group will use those values instead of their own.
-   */
-  privateLocations?: (string | PrivateLocation | PrivateLocationRef)[]
-
-  /**
-   * Tags for organizing and filtering checks.
-   *
-   * Checks in the group will behave as if they had the these tags set, though
-   * they may not be visible.
-   */
-  tags?: string[]
-
-  /**
-   * Determines how many checks are invoked concurrently when triggering a
-   * check group from CI/CD or through the API.
-   */
-  concurrency?: number
-
-  /**
-   * Optional fallback value for all checks belonging to the group. How often
-   * the check should run in minutes.
-   *
-   * Note that this settings is stored at the check level and is not persisted
-   * in the group.
-   */
-  frequency?: number | Frequency
-
-  /**
-   * When set, any environment variables defined here will be available to
-   * all checks in the group.
-   *
-   * If a global environment variable with the same name exists, the group
-   * environment variable will override it.
-   *
-   * If a check environment variable with the same name exists, it will
-   * override the group environment variable.
-   *
-   * See https://www.checklyhq.com/docs/groups/variables/#variable-hierarchy
-   * for more information.
-   */
-  environmentVariables?: EnvironmentVariable[]
-
-  /**
-   * List of alert channels to be alerted when checks in this group fail or
-   * recover.
-   */
-  alertChannels?: (AlertChannel | AlertChannelRef)[]
-
-  /**
-   * This optional setting can be used to provide CLI-level defaults for
-   * all browser checks in the group.
-   *
-   * Note that any settings defined here are stored at the check level and are
-   * not persisted in the group.
-   *
-   * Currently only the following settings have an effect:
-   * - {@link BrowserCheckConfig.frequency}
-   */
-  browserChecks?: BrowserCheckConfig,
-
-  /**
-   * This optional setting can be used to provide CLI-level defaults for
-   * all multi-step checks in the group.
-   *
-   * Note that any settings defined here are stored at the check level and are
-   * not persisted in the group.
-   *
-   * Currently only the following settings have an effect:
-   * - {@link MultiStepCheckConfig.frequency}
-   */
-  multiStepChecks?: MultiStepCheckConfig,
-
-  /**
-   * If set, all checks in the group will use the group's alert escalation
-   * policy.
-   *
-   * If not set, individual check settings are used.
-   */
-  alertEscalationPolicy?: AlertEscalation,
-
-  /**
-   * A valid piece of Node.js code to run in the setup phase of an API check
-   * in this group.
-   *
-   * @deprecated Use the "ApiCheck.setupScript" property instead and use
-   * common JS/TS code composition to add group specific setup routines.
-   */
-  localSetupScript?: string
-
-  /**
-   * A valid piece of Node.js code to run in the teardown phase of an API
-   * check in this group.
-   *
-   * @deprecated use the "ApiCheck.tearDownScript" property instead and use
-   * common JS/TS code composition to add group specific teardown routines.
-   */
-  localTearDownScript?: string
-
-  /**
-   * If set, the values provided here are merged with the values defined
-   * in individual API checks.
-   */
-  apiCheckDefaults?: ApiCheckDefaultConfig
-
-  /**
-   * Sets a retry policy for the group. Use {@link RetryStrategyBuilder} to
-   * create a retry policy.
-   *
-   * If set, all checks in the group use the group's retry strategy.
-   *
-   * If not set, individual check settings are used.
-   */
-  retryStrategy?: RetryStrategy
-
-  /**
-   * Determines whether the checks in the group should run on all selected
-   * locations in parallel or round-robin.
-   *
-   * When `true`, all checks in the group run in parallel regardless of their
-   * individual setting.
-   *
-   * When `false`, all checks in the group run in round-robin regardless of
-   * their individual setting.
-   *
-   * If not set, individual check settings are used.
-   *
-   * See https://www.checklyhq.com/docs/monitoring/global-locations/ to learn
-   * more about scheduling strategies.
-   */
-  runParallel?: boolean
-}
-
-/**
- * Creates a reference to an existing Check Group.
- *
- * References link existing resources to a project without managing them.
- */
-export class CheckGroupRef extends Construct {
-  constructor (logicalId: string, physicalId: string|number) {
-    super(CheckGroup.__checklyType, logicalId, physicalId, false)
-    Session.registerConstruct(this)
-  }
-
-  public getCheckDefaults (): CheckConfigDefaults {
-    // The only value CheckGroup.getCheckDefaults() returns is frequency,
-    // which exists purely for convenience, and only at CLI-level. It never
-    // gets sent to the backend. If references are being used, no access to
-    // this feature is required, and it can be ignored entirely.
-    return {}
-  }
-
-  public getBrowserCheckDefaults (): CheckConfigDefaults {
-    // See the comment for getCheckDefaults(), the same applies here.
-    return {}
-  }
-
-  public getMultiStepCheckDefaults (): CheckConfigDefaults {
-    // See the comment for getCheckDefaults(), the same applies here.
-    return {}
-  }
-
-  synthesize () {
-    return null
-  }
-}
-
-/**
- * Creates a Check Group
- *
- * @remarks
- *
- * This class make use of the Check Groups endpoints.
- */
-export class CheckGroup extends Construct {
-  name: string
-  activated?: boolean
-  muted?: boolean
-  doubleCheck?: boolean
-  runtimeId?: string
-  locations: Array<keyof Region>
-  privateLocations?: Array<string|PrivateLocation|PrivateLocationRef>
-  tags?: Array<string>
-  concurrency?: number
-  frequency?: number | Frequency
-  environmentVariables?: Array<EnvironmentVariable>
-  alertChannels?: Array<AlertChannel|AlertChannelRef>
-  localSetupScript?: string
-  localTearDownScript?: string
-  apiCheckDefaults: ApiCheckDefaultConfig
-  browserChecks?: BrowserCheckConfig
-  multiStepChecks?: MultiStepCheckConfig
-  retryStrategy?: RetryStrategy
-  runParallel?: boolean
-  alertSettings?: AlertEscalation
-  useGlobalAlertSettings?: boolean
-
-  static readonly __checklyType = 'check-group'
-
-  /**
-   * Constructs the CheckGroup instance
-   *
-   * @param logicalId unique project-scoped resource name identification
-   * @param props CheckGroup configuration properties
-   *
-   * {@link https://checklyhq.com/docs/cli/constructs-reference/#checkgroup Read more in the docs}
-   */
-  constructor (logicalId: string, props: CheckGroupProps) {
-    super(CheckGroup.__checklyType, logicalId)
-    this.name = props.name
-    this.activated = props.activated
-    this.muted = props.muted
-    this.doubleCheck = props.doubleCheck
-    this.tags = props.tags
-    this.runtimeId = props.runtimeId
-    this.locations = props.locations ?? []
-    this.privateLocations = props.privateLocations
-    this.concurrency = props.concurrency
-    // `frequency` is not a CheckGroup resource property. Not present in synthesize()
-    this.frequency = props.frequency
-    this.apiCheckDefaults = { ...defaultApiCheckDefaults, ...props.apiCheckDefaults }
-    this.alertSettings = props.alertEscalationPolicy
-    this.useGlobalAlertSettings = !this.alertSettings
-    this.environmentVariables = props.environmentVariables ?? []
-    this.alertChannels = props.alertChannels ?? []
-    this.localSetupScript = props.localSetupScript
-    this.localTearDownScript = props.localTearDownScript
-    this.retryStrategy = props.retryStrategy
-    this.runParallel = props.runParallel
-    // `browserChecks` is not a CheckGroup resource property. Not present in synthesize()
-    this.browserChecks = props.browserChecks
-    const fileAbsolutePath = Session.checkFileAbsolutePath!
-    if (props.browserChecks?.testMatch) {
-      this.__addChecks(fileAbsolutePath, props.browserChecks.testMatch, CheckTypes.BROWSER)
-    }
-    this.multiStepChecks = props.multiStepChecks
-    if (props.multiStepChecks?.testMatch) {
-      this.__addChecks(fileAbsolutePath, props.multiStepChecks.testMatch, CheckTypes.MULTI_STEP)
-    }
-    Session.registerConstruct(this)
-    this.__addSubscriptions()
-    this.__addPrivateLocationGroupAssignments()
-  }
-
-  async validate (diagnostics: Diagnostics): Promise<void> {
-    if (this.doubleCheck !== undefined) {
-      if (this.doubleCheck) {
-        diagnostics.add(new DeprecatedPropertyDiagnostic(
-          'doubleCheck',
-          new Error(
-            `To match the behavior of doubleCheck: true, please use the ` +
-            `following retryStategy instead:` +
-            `\n\n` +
-            `  RetryStrategyBuilder.fixedStrategy({\n` +
-            `    maxRetries: 1,\n` +
-            `    baseBackoffSeconds: 0,\n`+
-            `    maxDurationSeconds: 600,\n` +
-            `    sameRegion: false,\n` +
-            `  })`,
-          ),
-        ))
-      } else {
-        diagnostics.add(new DeprecatedPropertyDiagnostic(
-          'doubleCheck',
-          new Error(
-            `To match the behavior of doubleCheck: false, please use the ` +
-            `following retryStategy instead:` +
-            `\n\n` +
-            `  RetryStrategyBuilder.noRetries()`,
-          ),
-        ))
-      }
-    }
-
-    if (this.environmentVariables) {
-      this.environmentVariables.forEach(ev => {
-        // only empty string is checked because the KeyValuePair.value doesn't allow undefined or null.
-        if (ev.value === '') {
-          diagnostics.add(new InvalidPropertyValueDiagnostic(
-            'environmentVariables',
-            new Error(`Value must not be empty.`),
-          ))
-        }
-      })
-    }
-
-    if (this.localSetupScript) {
-      diagnostics.add(new DeprecatedPropertyDiagnostic(
-        'localSetupScript',
-        new Error(
-          `Use the setupScript property directly in your ApiChecks instead.`,
-        ),
-      ))
-    }
-
-    if (this.localTearDownScript) {
-      diagnostics.add(new DeprecatedPropertyDiagnostic(
-        'localTearDownScript',
-        new Error(
-          `Use the tearDownScript property directly in your ApiChecks instead.`,
-        ),
-      ))
-    }
-  }
-
-  static fromId (id: number) {
-    return new CheckGroupRef(`check-group-${id}`, id)
-  }
-
-  private __addChecks (
-    fileAbsolutePath: string,
-    testMatch: string|string[],
-    checkType: typeof CheckTypes.BROWSER | typeof CheckTypes.MULTI_STEP,
-  ) {
-    const parent = path.dirname(fileAbsolutePath)
-    const matched = glob.sync(testMatch, { nodir: true, cwd: parent })
-    for (const match of matched) {
-      const filepath = path.join(parent, match)
-      const props = {
-        group: this,
-        name: match,
-        code: {
-          entrypoint: filepath,
-        },
-        // the browserChecks props inherited from the group are applied in BrowserCheck.constructor()
-      }
-      const checkLogicalId = pathToPosix(path.relative(Session.basePath!, filepath))
-      if (checkType === CheckTypes.BROWSER) {
-        new BrowserCheck(checkLogicalId, props)
-      } else {
-        new MultiStepCheck(checkLogicalId, props)
-      }
-    }
-  }
-
-  private __addSubscriptions () {
-    if (!this.alertChannels) {
-      return
-    }
-    for (const alertChannel of this.alertChannels) {
-      const subscription = new AlertChannelSubscription(`check-group-alert-channel-subscription#${this.logicalId}#${alertChannel.logicalId}`, {
-        alertChannelId: Ref.from(alertChannel.logicalId),
-        groupId: Ref.from(this.logicalId),
-        activated: true,
-      })
-    }
-  }
-
-  private __addPrivateLocationGroupAssignments () {
-    if (!this.privateLocations) {
-      return
-    }
-    for (const privateLocation of this.privateLocations) {
-      // slugName strings are processed in loadAllPrivateLocations()
-      if (typeof privateLocation === 'string') {
-        continue
-      }
-
-      // use private location assignment for instances
-      const assignment = new PrivateLocationGroupAssignment(`private-location-group-assignment#${this.logicalId}#${privateLocation.logicalId}`, {
-        privateLocationId: Ref.from(privateLocation.logicalId),
-        groupId: Ref.from(this.logicalId),
-      })
-    }
-  }
-
-  public getCheckDefaults (): CheckConfigDefaults {
-    // TODO: investigate if make sense to add all other check's properties
-    return {
-      frequency: this.frequency,
-    }
-  }
-
-  public getBrowserCheckDefaults (): CheckConfigDefaults {
-    // TODO: investigate if make sense to add all other browser-check's properties
-    return {
-      frequency: this.browserChecks?.frequency,
-    }
-  }
-
-  public getMultiStepCheckDefaults (): CheckConfigDefaults {
-    return {
-      frequency: this.multiStepChecks?.frequency,
-    }
-  }
-
-  synthesize () {
-    return {
-      name: this.name,
-      activated: this.activated,
-      muted: this.muted,
-      tags: this.tags,
-      locations: this.locations,
-      runtimeId: this.runtimeId,
-
-      // private-location instances are assigned with loadAllPrivateLocations()
-      privateLocations: undefined,
-
-      concurrency: this.concurrency,
-      localSetupScript: this.localSetupScript,
-      localTearDownScript: this.localTearDownScript,
-      apiCheckDefaults: this.apiCheckDefaults,
-      environmentVariables: this.environmentVariables,
-      // The backend doesn't actually support the `NO_RETRIES` type, it uses `null` instead.
-      retryStrategy: this.retryStrategy?.type === 'NO_RETRIES'
-        ? null
-        : this.retryStrategy,
-      // When `retryStrategy: NO_RETRIES` and `doubleCheck: undefined`, we want to let the user disable all retries.
-      // The backend has a Joi default of `doubleCheck: true`, though, so we need special handling for this case.
-      doubleCheck: this.doubleCheck === undefined && this.retryStrategy?.type === 'NO_RETRIES'
-        ? false
-        : this.doubleCheck,
-      runParallel: this.runParallel,
-      alertSettings: this.alertSettings,
-      useGlobalAlertSettings: this.useGlobalAlertSettings,
-    }
-  }
-}
